+++ src/index.ts
@@ -1,6 +1,7 @@
 export { closest } from "./closest";
 export { setAttribute } from './set_attribute';
 export { getAttribute } from './get_attribute';
+export { hasAttribute } from './has_attribute';
 export { querySelector } from "./query_selector";
 export { createElement } from "./create_element";
 export { createTextNode } from "./create_text_node";